refactor: 优化设计模块的类型导入和结构，减少重复定义
